+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `replace()` method
 - Added `min()` and `max()` methods
 - Added `sum()` method
+- Added `indexOf()` and `lastIndexOf()` methods
 - Added `toLowerCase()`, `toUpperCase()` and `trim()` methods
 - Added math methods `abs()`, `acos()`, `acosh()`, `asin()`, `asinh()`, `atan()`, `atan2()`, `atanh()`, `cbrt()`, `ceil()`, `clz32()`, `cos()`, `cosh()`, `exp()`, `expm1()`, `floor()`, `fround()`, `hypot()`, `imul()`, `log()`, `log10()`, `log1p()`, `log2()`, `pow()`, `round()`, `sign()`, `sin()`, `sinh()`, `sqrt()`, `tan()`, `tanh()` and `trunc()`. All the methods works the same as static methods of [`Math`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Math) in JavaScript
 - Changed the comparator function grammar to allow any expression in the definition, not just a query chain. This eliminates the need for parentheses in many cases. For example, `a + b desc` is now a valid definition, whereas previously it required parentheses: `(a + b) desc` (since evaluated as `a + (b desc)`).